+++ e2e/testcases/correction-birth/correct-birth-record-6.spec.ts
@@ -2,7 +2,6 @@ import { expect, test, type Page } from '@playwright/test'
 import {
   assignRecord,
   auditRecord,
-  createPIN,
   expectAddress,
   expectOutboxToBeEmpty,
   formatDateTo_ddMMMMyyyy,
