+++ e2e/testcases/v2-print-certificate/death/36.10-validate-certify-record-page.spec.ts
@@ -11,6 +11,7 @@ import {
   selectRequesterType
 } from '../death/helpers'
 import { expectInUrl } from '../../../v2-utils'
+import { printAndExpectPopup } from '../birth/helpers'
 
 test.describe.serial('10.0 Validate "Review" page', () => {
   let page: Page
@@ -65,16 +66,7 @@ test.describe.serial('10.0 Validate "Review" page', () => {
 
   test('10.3 On click print button, user will navigate to a new tab from where user can download PDF', async () => {
     await page.getByRole('button', { name: 'Yes, print certificate' }).click()
-
-    const popupPromise = page.waitForEvent('popup')
-    await page.getByRole('button', { name: 'Print', exact: true }).click()
-    const popup = await popupPromise
-    const downloadPromise = popup.waitForEvent('download')
-    const download = await downloadPromise
-
-    // Check that the popup URL contains PDF content
-    await expect(popup.url()).toBe('about:blank')
-    await expect(download.suggestedFilename()).toMatch(/^.*\.pdf$/)
+    await printAndExpectPopup(page)
 
     await expectInUrl(page, `/events/overview/${eventId}`)
   })
